refactor : movieRatingService 폴더구조 변경
- event utils 함수에서 분리
